Move npm commands to shell scripts
