+++ test/integration/errors.js
@@ -5,8 +5,8 @@ const pify = require('pify')
 const shell = require('shelljs');
 const ganache = require('ganache-cli')
 
-const verify = require('../../util/verifiers')
-const mock = require('../../util/integration');
+const verify = require('./../util/verifiers')
+const mock = require('./../util/integration');
 
 // =======
 // Errors
